feat(redux-saga): move all effects to CommonEffects
