+++ android/android-livedata-saga/src/test/java/org/swiften/redux/android/saga/rx/livedata/LiveDataEffectTest.kt
@@ -8,7 +8,9 @@ package org.swiften.redux.android.saga.rx.livedata
 import androidx.arch.core.executor.testing.InstantTaskExecutorRule
 import androidx.lifecycle.MutableLiveData
 import kotlinx.coroutines.delay
+import kotlinx.coroutines.isActive
 import kotlinx.coroutines.runBlocking
+import kotlinx.coroutines.withTimeoutOrNull
 import org.junit.Assert.assertEquals
 import org.junit.Rule
 import org.junit.Test
@@ -17,7 +19,7 @@ import org.junit.runner.RunWith
 import org.robolectric.RobolectricTestRunner
 import org.robolectric.annotation.Config
 import org.swiften.redux.android.saga.rx.livedata.LiveDataEffects.takeLiveData
-import org.swiften.redux.saga.common.SagaEffects.await
+import org.swiften.redux.saga.common.CommonEffects.await
 import org.swiften.redux.saga.common.SagaInput
 import org.swiften.redux.saga.common.SagaMonitor
 import org.swiften.redux.saga.common.flatMap
@@ -29,6 +31,8 @@ import java.util.Collections
 class LiveDataEffectTest {
   @get:Rule
   val rule: TestRule = InstantTaskExecutorRule()
+
+  private val timeout: Long = 10000
 
   @Test
   fun test_takeLiveDataEffect_shouldStreamLiveData() {
@@ -46,7 +50,9 @@ class LiveDataEffectTest {
     (0 until 5).forEach { data.value = it }
 
     runBlocking {
-      delay(1500)
+      withTimeoutOrNull(this@LiveDataEffectTest.timeout) {
+        while (finalValues.size != 5 && this.isActive) { delay(500) }; Unit
+      }
 
       // Then
       assertEquals(finalValues.sorted(), arrayListOf(0, 1, 2, 3, 4))
