+++ common/common-saga/src/main/kotlin/org/swiften/redux/saga/common/CommonExtension.kt
@@ -7,6 +7,17 @@ package org.swiften.redux.saga.common
 
 /** Created by haipham on 2019/01/26/1 */
 /**
+ * Invoke a [DebounceEffect] on [this].
+ * @receiver See [DebounceEffect.source].
+ * @param R The result emission type.
+ * @param millis See [DebounceEffect.millis].
+ * @return A [SagaEffect] instance.
+ */
+fun <R> SagaEffect<R>.debounce(millis: Long): SagaEffect<R> where R : Any {
+  return (CommonEffects.debounce<R>(millis))(this)
+}
+
+/**
  * Invoke a [FlatMapEffect] on the current [ISagaEffect] with [FlatMapEffect.Mode.EVERY].
  * @receiver See [FlatMapEffect.source].
  * @param P The source emission type.
