+++ docs/org.swiften.redux.saga.rx/-saga-effects/await.md
@@ -2,7 +2,7 @@
 
 # await
 
-`fun <R : `[`Any`](https://kotlinlang.org/api/latest/jvm/stdlib/kotlin/-any/index.html)`> await(creator: `[`IAwaitCreator`](../-i-await-creator.md)`<`[`R`](await.md#R)`>): `[`SingleSagaEffect`](../../org.swiften.redux.saga.common/-single-saga-effect/index.md)`<`[`R`](await.md#R)`>` [(source)](https://github.com/protoman92/KotlinRedux/tree/master/common/common-rx-saga/src/main/kotlin/org/swiften/redux/saga/rx/SagaEffects.kt#L32)
+`fun <R : `[`Any`](https://kotlinlang.org/api/latest/jvm/stdlib/kotlin/-any/index.html)`> await(creator: `[`IAwaitCreator`](../-i-await-creator.md)`<`[`R`](await.md#R)`>): `[`SingleSagaEffect`](../../org.swiften.redux.saga.common/-single-saga-effect/index.md)`<`[`R`](await.md#R)`>` [(source)](https://github.com/protoman92/KotlinRedux/tree/master/common/common-rx-saga/src/main/kotlin/org/swiften/redux/saga/rx/CommonEffects.kt#L32)
 
 Create an [AwaitEffect](../-await-effect/index.md) instance.
