+++ docs/org.swiften.redux.saga.rx/-saga-effects/take-every-action-for-keys.md
@@ -2,7 +2,7 @@
 
 # takeEveryActionForKeys
 
-`fun <R : `[`Any`](https://kotlinlang.org/api/latest/jvm/stdlib/kotlin/-any/index.html)`> takeEveryActionForKeys(actionKeys: `[`Set`](https://kotlinlang.org/api/latest/jvm/stdlib/kotlin.collections/-set/index.html)`<`[`String`](https://kotlinlang.org/api/latest/jvm/stdlib/kotlin/-string/index.html)`>, creator: (`[`IReduxActionWithKey`](../../org.swiften.redux.core/-i-redux-action-with-key/index.md)`) -> `[`ISagaEffect`](../../org.swiften.redux.saga.common/-i-saga-effect.md)`<`[`R`](take-every-action-for-keys.md#R)`>): `[`TakeActionEffect`](../../org.swiften.redux.saga.common/-take-action-effect/index.md)`<`[`IReduxActionWithKey`](../../org.swiften.redux.core/-i-redux-action-with-key/index.md)`, `[`IReduxActionWithKey`](../../org.swiften.redux.core/-i-redux-action-with-key/index.md)`, `[`R`](take-every-action-for-keys.md#R)`>` [(source)](https://github.com/protoman92/KotlinRedux/tree/master/common/common-rx-saga/src/main/kotlin/org/swiften/redux/saga/rx/SagaEffects.kt#L200)
+`fun <R : `[`Any`](https://kotlinlang.org/api/latest/jvm/stdlib/kotlin/-any/index.html)`> takeEveryActionForKeys(actionKeys: `[`Set`](https://kotlinlang.org/api/latest/jvm/stdlib/kotlin.collections/-set/index.html)`<`[`String`](https://kotlinlang.org/api/latest/jvm/stdlib/kotlin/-string/index.html)`>, creator: (`[`IReduxActionWithKey`](../../org.swiften.redux.core/-i-redux-action-with-key/index.md)`) -> `[`ISagaEffect`](../../org.swiften.redux.saga.common/-i-saga-effect.md)`<`[`R`](take-every-action-for-keys.md#R)`>): `[`TakeActionEffect`](../../org.swiften.redux.saga.common/-take-action-effect/index.md)`<`[`IReduxActionWithKey`](../../org.swiften.redux.core/-i-redux-action-with-key/index.md)`, `[`IReduxActionWithKey`](../../org.swiften.redux.core/-i-redux-action-with-key/index.md)`, `[`R`](take-every-action-for-keys.md#R)`>` [(source)](https://github.com/protoman92/KotlinRedux/tree/master/common/common-rx-saga/src/main/kotlin/org/swiften/redux/saga/rx/CommonEffects.kt#L200)
 
 Instead of specifying the action type, check if [IReduxAction](../../org.swiften.redux.core/-i-redux-action.md) instances that pass through
 the pipeline conform to [IReduxActionWithKey](../../org.swiften.redux.core/-i-redux-action-with-key/index.md), and whether their [IReduxActionWithKey.key](../../org.swiften.redux.core/-i-redux-action-with-key/key.md)
