+++ docs/org.swiften.redux.saga.rx/-saga-effects/take-every-state.md
@@ -2,7 +2,7 @@
 
 # takeEveryState
 
-`fun <State : `[`Any`](https://kotlinlang.org/api/latest/jvm/stdlib/kotlin/-any/index.html)`, R : `[`Any`](https://kotlinlang.org/api/latest/jvm/stdlib/kotlin/-any/index.html)`> takeEveryState(cls: `[`Class`](http://docs.oracle.com/javase/6/docs/api/java/lang/Class.html)`<`[`State`](take-every-state.md#State)`>, creator: (`[`State`](take-every-state.md#State)`) -> `[`ISagaEffect`](../../org.swiften.redux.saga.common/-i-saga-effect.md)`<`[`R`](take-every-state.md#R)`>): `[`TakeStateEffect`](../../org.swiften.redux.saga.common/-take-state-effect/index.md)`<`[`State`](take-every-state.md#State)`, `[`R`](take-every-state.md#R)`>` [(source)](https://github.com/protoman92/KotlinRedux/tree/master/common/common-rx-saga/src/main/kotlin/org/swiften/redux/saga/rx/SagaEffects.kt#L273)
+`fun <State : `[`Any`](https://kotlinlang.org/api/latest/jvm/stdlib/kotlin/-any/index.html)`, R : `[`Any`](https://kotlinlang.org/api/latest/jvm/stdlib/kotlin/-any/index.html)`> takeEveryState(cls: `[`Class`](http://docs.oracle.com/javase/6/docs/api/java/lang/Class.html)`<`[`State`](take-every-state.md#State)`>, creator: (`[`State`](take-every-state.md#State)`) -> `[`ISagaEffect`](../../org.swiften.redux.saga.common/-i-saga-effect.md)`<`[`R`](take-every-state.md#R)`>): `[`TakeStateEffect`](../../org.swiften.redux.saga.common/-take-state-effect/index.md)`<`[`State`](take-every-state.md#State)`, `[`R`](take-every-state.md#R)`>` [(source)](https://github.com/protoman92/KotlinRedux/tree/master/common/common-rx-saga/src/main/kotlin/org/swiften/redux/saga/rx/CommonEffects.kt#L273)
 
 Create a [TakeEveryStateEffect](../-take-every-state-effect/index.md) instance.
 
@@ -15,7 +15,7 @@ Create a [TakeEveryStateEffect](../-take-every-state-effect/index.md) instance.
 **Return**
 A [TakeStateEffect](../../org.swiften.redux.saga.common/-take-state-effect/index.md) instance.
 
-`fun <State : `[`Any`](https://kotlinlang.org/api/latest/jvm/stdlib/kotlin/-any/index.html)`, R : `[`Any`](https://kotlinlang.org/api/latest/jvm/stdlib/kotlin/-any/index.html)`> takeEveryState(cls: `[`KClass`](https://kotlinlang.org/api/latest/jvm/stdlib/kotlin.reflect/-k-class/index.html)`<`[`State`](take-every-state.md#State)`>, creator: (`[`State`](take-every-state.md#State)`) -> `[`ISagaEffect`](../../org.swiften.redux.saga.common/-i-saga-effect.md)`<`[`R`](take-every-state.md#R)`>): `[`TakeStateEffect`](../../org.swiften.redux.saga.common/-take-state-effect/index.md)`<`[`State`](take-every-state.md#State)`, `[`R`](take-every-state.md#R)`>` [(source)](https://github.com/protoman92/KotlinRedux/tree/master/common/common-rx-saga/src/main/kotlin/org/swiften/redux/saga/rx/SagaEffects.kt#L287)
+`fun <State : `[`Any`](https://kotlinlang.org/api/latest/jvm/stdlib/kotlin/-any/index.html)`, R : `[`Any`](https://kotlinlang.org/api/latest/jvm/stdlib/kotlin/-any/index.html)`> takeEveryState(cls: `[`KClass`](https://kotlinlang.org/api/latest/jvm/stdlib/kotlin.reflect/-k-class/index.html)`<`[`State`](take-every-state.md#State)`>, creator: (`[`State`](take-every-state.md#State)`) -> `[`ISagaEffect`](../../org.swiften.redux.saga.common/-i-saga-effect.md)`<`[`R`](take-every-state.md#R)`>): `[`TakeStateEffect`](../../org.swiften.redux.saga.common/-take-state-effect/index.md)`<`[`State`](take-every-state.md#State)`, `[`R`](take-every-state.md#R)`>` [(source)](https://github.com/protoman92/KotlinRedux/tree/master/common/common-rx-saga/src/main/kotlin/org/swiften/redux/saga/rx/CommonEffects.kt#L287)
 
 Same as [takeEveryState](./take-every-state.md), but use [KClass](https://kotlinlang.org/api/latest/jvm/stdlib/kotlin.reflect/-k-class/index.html) instead of [Class](http://docs.oracle.com/javase/6/docs/api/java/lang/Class.html).
