+++ docs/org.swiften.redux.saga.rx/-saga-effects/take-latest-action.md
@@ -2,7 +2,7 @@
 
 # takeLatestAction
 
-`fun <Action : `[`IReduxAction`](../../org.swiften.redux.core/-i-redux-action.md)`, P : `[`Any`](https://kotlinlang.org/api/latest/jvm/stdlib/kotlin/-any/index.html)`, R : `[`Any`](https://kotlinlang.org/api/latest/jvm/stdlib/kotlin/-any/index.html)`> takeLatestAction(cls: `[`Class`](http://docs.oracle.com/javase/6/docs/api/java/lang/Class.html)`<`[`Action`](take-latest-action.md#Action)`>, extractor: (`[`Action`](take-latest-action.md#Action)`) -> `[`P`](take-latest-action.md#P)`?, creator: (`[`P`](take-latest-action.md#P)`) -> `[`ISagaEffect`](../../org.swiften.redux.saga.common/-i-saga-effect.md)`<`[`R`](take-latest-action.md#R)`>): `[`TakeActionEffect`](../../org.swiften.redux.saga.common/-take-action-effect/index.md)`<`[`Action`](take-latest-action.md#Action)`, `[`P`](take-latest-action.md#P)`, `[`R`](take-latest-action.md#R)`>` [(source)](https://github.com/protoman92/KotlinRedux/tree/master/common/common-rx-saga/src/main/kotlin/org/swiften/redux/saga/rx/SagaEffects.kt#L219)
+`fun <Action : `[`IReduxAction`](../../org.swiften.redux.core/-i-redux-action.md)`, P : `[`Any`](https://kotlinlang.org/api/latest/jvm/stdlib/kotlin/-any/index.html)`, R : `[`Any`](https://kotlinlang.org/api/latest/jvm/stdlib/kotlin/-any/index.html)`> takeLatestAction(cls: `[`Class`](http://docs.oracle.com/javase/6/docs/api/java/lang/Class.html)`<`[`Action`](take-latest-action.md#Action)`>, extractor: (`[`Action`](take-latest-action.md#Action)`) -> `[`P`](take-latest-action.md#P)`?, creator: (`[`P`](take-latest-action.md#P)`) -> `[`ISagaEffect`](../../org.swiften.redux.saga.common/-i-saga-effect.md)`<`[`R`](take-latest-action.md#R)`>): `[`TakeActionEffect`](../../org.swiften.redux.saga.common/-take-action-effect/index.md)`<`[`Action`](take-latest-action.md#Action)`, `[`P`](take-latest-action.md#P)`, `[`R`](take-latest-action.md#R)`>` [(source)](https://github.com/protoman92/KotlinRedux/tree/master/common/common-rx-saga/src/main/kotlin/org/swiften/redux/saga/rx/CommonEffects.kt#L219)
 
 Create a [TakeLatestActionEffect](../-take-latest-action-effect/index.md) instance.
 
@@ -18,7 +18,7 @@ Create a [TakeLatestActionEffect](../-take-latest-action-effect/index.md) instan
 
 `extractor` - Function that extracts [P](take-latest-action.md#P) from [IReduxAction](../../org.swiften.redux.core/-i-redux-action.md).
 
-`creator` - Function that creates [ISagaEffect](../../org.swiften.redux.saga.common/-i-saga-effect.md) from [P](take-latest-action.md#P).`fun <Action : `[`IReduxAction`](../../org.swiften.redux.core/-i-redux-action.md)`, P : `[`Any`](https://kotlinlang.org/api/latest/jvm/stdlib/kotlin/-any/index.html)`, R : `[`Any`](https://kotlinlang.org/api/latest/jvm/stdlib/kotlin/-any/index.html)`> takeLatestAction(cls: `[`KClass`](https://kotlinlang.org/api/latest/jvm/stdlib/kotlin.reflect/-k-class/index.html)`<`[`Action`](take-latest-action.md#Action)`>, extractor: (`[`Action`](take-latest-action.md#Action)`) -> `[`P`](take-latest-action.md#P)`?, creator: (`[`P`](take-latest-action.md#P)`) -> `[`ISagaEffect`](../../org.swiften.redux.saga.common/-i-saga-effect.md)`<`[`R`](take-latest-action.md#R)`>): `[`TakeActionEffect`](../../org.swiften.redux.saga.common/-take-action-effect/index.md)`<`[`Action`](take-latest-action.md#Action)`, `[`P`](take-latest-action.md#P)`, `[`R`](take-latest-action.md#R)`>` [(source)](https://github.com/protoman92/KotlinRedux/tree/master/common/common-rx-saga/src/main/kotlin/org/swiften/redux/saga/rx/SagaEffects.kt#L237)
+`creator` - Function that creates [ISagaEffect](../../org.swiften.redux.saga.common/-i-saga-effect.md) from [P](take-latest-action.md#P).`fun <Action : `[`IReduxAction`](../../org.swiften.redux.core/-i-redux-action.md)`, P : `[`Any`](https://kotlinlang.org/api/latest/jvm/stdlib/kotlin/-any/index.html)`, R : `[`Any`](https://kotlinlang.org/api/latest/jvm/stdlib/kotlin/-any/index.html)`> takeLatestAction(cls: `[`KClass`](https://kotlinlang.org/api/latest/jvm/stdlib/kotlin.reflect/-k-class/index.html)`<`[`Action`](take-latest-action.md#Action)`>, extractor: (`[`Action`](take-latest-action.md#Action)`) -> `[`P`](take-latest-action.md#P)`?, creator: (`[`P`](take-latest-action.md#P)`) -> `[`ISagaEffect`](../../org.swiften.redux.saga.common/-i-saga-effect.md)`<`[`R`](take-latest-action.md#R)`>): `[`TakeActionEffect`](../../org.swiften.redux.saga.common/-take-action-effect/index.md)`<`[`Action`](take-latest-action.md#Action)`, `[`P`](take-latest-action.md#P)`, `[`R`](take-latest-action.md#R)`>` [(source)](https://github.com/protoman92/KotlinRedux/tree/master/common/common-rx-saga/src/main/kotlin/org/swiften/redux/saga/rx/CommonEffects.kt#L237)
 
 Convenience function to invoke [takeLatestActionForKeys](take-latest-action-for-keys.md) using [KClass](https://kotlinlang.org/api/latest/jvm/stdlib/kotlin.reflect/-k-class/index.html) instead of [Class](http://docs.oracle.com/javase/6/docs/api/java/lang/Class.html).
