+++ docs/org.swiften.redux.saga.rx/-saga-effects/take-latest-state.md
@@ -2,7 +2,7 @@
 
 # takeLatestState
 
-`fun <State : `[`Any`](https://kotlinlang.org/api/latest/jvm/stdlib/kotlin/-any/index.html)`, R : `[`Any`](https://kotlinlang.org/api/latest/jvm/stdlib/kotlin/-any/index.html)`> takeLatestState(cls: `[`Class`](http://docs.oracle.com/javase/6/docs/api/java/lang/Class.html)`<`[`State`](take-latest-state.md#State)`>, creator: (`[`State`](take-latest-state.md#State)`) -> `[`ISagaEffect`](../../org.swiften.redux.saga.common/-i-saga-effect.md)`<`[`R`](take-latest-state.md#R)`>): `[`TakeStateEffect`](../../org.swiften.redux.saga.common/-take-state-effect/index.md)`<`[`State`](take-latest-state.md#State)`, `[`R`](take-latest-state.md#R)`>` [(source)](https://github.com/protoman92/KotlinRedux/tree/master/common/common-rx-saga/src/main/kotlin/org/swiften/redux/saga/rx/SagaEffects.kt#L301)
+`fun <State : `[`Any`](https://kotlinlang.org/api/latest/jvm/stdlib/kotlin/-any/index.html)`, R : `[`Any`](https://kotlinlang.org/api/latest/jvm/stdlib/kotlin/-any/index.html)`> takeLatestState(cls: `[`Class`](http://docs.oracle.com/javase/6/docs/api/java/lang/Class.html)`<`[`State`](take-latest-state.md#State)`>, creator: (`[`State`](take-latest-state.md#State)`) -> `[`ISagaEffect`](../../org.swiften.redux.saga.common/-i-saga-effect.md)`<`[`R`](take-latest-state.md#R)`>): `[`TakeStateEffect`](../../org.swiften.redux.saga.common/-take-state-effect/index.md)`<`[`State`](take-latest-state.md#State)`, `[`R`](take-latest-state.md#R)`>` [(source)](https://github.com/protoman92/KotlinRedux/tree/master/common/common-rx-saga/src/main/kotlin/org/swiften/redux/saga/rx/CommonEffects.kt#L301)
 
 Create a [TakeLatestStateEffect](../-take-latest-state-effect/index.md) instance.
 
@@ -15,7 +15,7 @@ Create a [TakeLatestStateEffect](../-take-latest-state-effect/index.md) instance
 **Return**
 A [TakeStateEffect](../../org.swiften.redux.saga.common/-take-state-effect/index.md) instance.
 
-`fun <State : `[`Any`](https://kotlinlang.org/api/latest/jvm/stdlib/kotlin/-any/index.html)`, R : `[`Any`](https://kotlinlang.org/api/latest/jvm/stdlib/kotlin/-any/index.html)`> takeLatestState(cls: `[`KClass`](https://kotlinlang.org/api/latest/jvm/stdlib/kotlin.reflect/-k-class/index.html)`<`[`State`](take-latest-state.md#State)`>, creator: (`[`State`](take-latest-state.md#State)`) -> `[`ISagaEffect`](../../org.swiften.redux.saga.common/-i-saga-effect.md)`<`[`R`](take-latest-state.md#R)`>): `[`TakeStateEffect`](../../org.swiften.redux.saga.common/-take-state-effect/index.md)`<`[`State`](take-latest-state.md#State)`, `[`R`](take-latest-state.md#R)`>` [(source)](https://github.com/protoman92/KotlinRedux/tree/master/common/common-rx-saga/src/main/kotlin/org/swiften/redux/saga/rx/SagaEffects.kt#L315)
+`fun <State : `[`Any`](https://kotlinlang.org/api/latest/jvm/stdlib/kotlin/-any/index.html)`, R : `[`Any`](https://kotlinlang.org/api/latest/jvm/stdlib/kotlin/-any/index.html)`> takeLatestState(cls: `[`KClass`](https://kotlinlang.org/api/latest/jvm/stdlib/kotlin.reflect/-k-class/index.html)`<`[`State`](take-latest-state.md#State)`>, creator: (`[`State`](take-latest-state.md#State)`) -> `[`ISagaEffect`](../../org.swiften.redux.saga.common/-i-saga-effect.md)`<`[`R`](take-latest-state.md#R)`>): `[`TakeStateEffect`](../../org.swiften.redux.saga.common/-take-state-effect/index.md)`<`[`State`](take-latest-state.md#State)`, `[`R`](take-latest-state.md#R)`>` [(source)](https://github.com/protoman92/KotlinRedux/tree/master/common/common-rx-saga/src/main/kotlin/org/swiften/redux/saga/rx/CommonEffects.kt#L315)
 
 Same as [takeLatestState](./take-latest-state.md), but use [KClass](https://kotlinlang.org/api/latest/jvm/stdlib/kotlin.reflect/-k-class/index.html) instead of [Class](http://docs.oracle.com/javase/6/docs/api/java/lang/Class.html).
