+++ docs/org.swiften.redux.saga.rx/-saga-effects/then-no-matter-what.md
@@ -2,7 +2,7 @@
 
 # thenNoMatterWhat
 
-`fun <R : `[`Any`](https://kotlinlang.org/api/latest/jvm/stdlib/kotlin/-any/index.html)`, R2 : `[`Any`](https://kotlinlang.org/api/latest/jvm/stdlib/kotlin/-any/index.html)`> thenNoMatterWhat(defaultValue: `[`R2`](then-no-matter-what.md#R2)`): `[`ISagaEffectTransformer`](../../org.swiften.redux.saga.common/-i-saga-effect-transformer.md)`<`[`R`](then-no-matter-what.md#R)`, `[`R2`](then-no-matter-what.md#R2)`>` [(source)](https://github.com/protoman92/KotlinRedux/tree/master/common/common-rx-saga/src/main/kotlin/org/swiften/redux/saga/rx/SagaEffects.kt#L147)
+`fun <R : `[`Any`](https://kotlinlang.org/api/latest/jvm/stdlib/kotlin/-any/index.html)`, R2 : `[`Any`](https://kotlinlang.org/api/latest/jvm/stdlib/kotlin/-any/index.html)`> thenNoMatterWhat(defaultValue: `[`R2`](then-no-matter-what.md#R2)`): `[`ISagaEffectTransformer`](../../org.swiften.redux.saga.common/-i-saga-effect-transformer.md)`<`[`R`](then-no-matter-what.md#R)`, `[`R2`](then-no-matter-what.md#R2)`>` [(source)](https://github.com/protoman92/KotlinRedux/tree/master/common/common-rx-saga/src/main/kotlin/org/swiften/redux/saga/rx/CommonEffects.kt#L147)
 
 Convenience method for [CommonEffects.thenNoMatterWhat](../../org.swiften.redux.saga.common/-common-effects/then-no-matter-what.md) to switch to [defaultValue](then-no-matter-what.md#org.swiften.redux.saga.rx.SagaEffects$thenNoMatterWhat(org.swiften.redux.saga.rx.SagaEffects.thenNoMatterWhat.R2)/defaultValue).
